requirements.txt
requests
pandas
pdfminer.six
pdfplumber
